+++ lib/easy_auth/models/account.rb
@@ -4,5 +4,6 @@ module EasyAuth::Models::Account
   included do
     # Relationships
     has_many :identities, :class_name => 'Identity', :as => :account, :dependent => :destroy, :autosave => true
+    accepts_nested_attributes_for :identities
   end
 end