+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Reverted array identities
 * Updated Rails to ~> 4.0.0
+* Identities are destroy dependent of parent