+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 * Reverted array identities
 * Updated Rails to ~> 4.0.0
 * Identities are destroy dependent of parent
+* Identities are always autosave