+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * Updated Rails to ~> 4.0.0
 * Identities are destroy dependent of parent
 * Identities are always autosave
+* Identities will require validation for parent to save