+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 * Identities will require validation for parent to save
 * Identities require presence of token
 * Accounts accept nested attributes for identities
+* SessionsController#create will rescue DoubleRenderErrors